Cast sig_atomic_t global_interrupt to int in os_signal_interrupt()

+++ cmdline/mingwapp.c
@@ -24,7 +24,7 @@ volatile sig_atomic_t global_interrupt = 0;
 
 int os_signal_interrupt(void)
 {
-	return global_interrupt;
+	return (int)global_interrupt;
 }
 
 void app_signal_handler(int signum)
